[배환웅] 포인트 교환 로직 수정

+++ PointGame/Assets/Script/TKManager.cs
@@ -233,7 +233,7 @@ public class TKManager : MonoBehaviour
             return 0;
 
         int tempPoint = point;
-        int tempCash = (tempPoint / CommonData.PointToCashChange) * CommonData.PointToCashChangeValue;
+        int tempCash = (int)((float)(tempPoint / CommonData.PointToCashChange) * CommonData.PointToCashChangeValue);
 
         int count = 0;
         int bonusCount = 0;
